src/requirements.txt
prefect
minio
boto3
clickhouse-connect
requests
pandas
pydantic
python-telegram-bot
python-dotenv